@@ -1,74 +1,99 @@
 {
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "^5.13.0",
-    "@sentry/cli": "^1.52.3",
-    "@sentry/electron": "^2.0.4",
+    "@angular/animations": "^12.0.0",
+    "@angular/common": "^12.0.0",
+    "@angular/compiler": "^12.0.0",
+    "@angular/core": "^12.0.0",
+    "@angular/forms": "^12.0.0",
+    "@angular/platform-browser": "^12.0.0",
+    "@angular/platform-browser-dynamic": "^12.0.0",
+    "@fortawesome/fontawesome-free": "^5.15.3",
+    "@ng-bootstrap/ng-bootstrap": "^10.0.0",
+    "@sentry/cli": "^1.67.2",
+    "@sentry/electron": "^2.5.1",
+    "@tabby-gang/to-string-loader": "^1.1.7-beta.2",
     "@types/electron-config": "^3.2.2",
     "@types/electron-debug": "^2.1.0",
-    "@types/js-yaml": "^3.12.4",
-    "@types/node": "12.7.12",
-    "@types/webpack-env": "^1.15.2",
-    "@typescript-eslint/eslint-plugin": "^2.26.0",
-    "@typescript-eslint/parser": "^3.8.0",
+    "@types/fs-extra": "^9.0.12",
+    "@types/js-yaml": "^4.0.2",
+    "@types/node": "16.0.1",
+    "@types/sortablejs": "^1.10.7",
+    "@types/webpack-env": "^1.16.2",
+    "@typescript-eslint/eslint-plugin": "^4.28.3",
+    "@typescript-eslint/parser": "^4.28.5",
     "apply-loader": "2.0.0",
-    "awesome-typescript-loader": "^5.0.0",
-    "core-js": "^3.6.5",
-    "cross-env": "7.0.2",
-    "css-loader": "3.4.2",
-    "electron": "^8.5.2",
-    "electron-builder": "22.6.1",
+    "axios": "^0.21.1",
+    "clone-deep": "^4.0.1",
+    "compare-versions": "^3.6.0",
+    "core-js": "^3.15.2",
+    "cross-env": "7.0.3",
+    "css-loader": "5.2.6",
+    "electron": "13.1.7",
+    "electron-builder": "22.10.5",
     "electron-download": "^4.1.1",
-    "electron-installer-snap": "^5.0.0",
+    "electron-installer-snap": "^5.1.0",
     "electron-notarize": "^1.0.0",
-    "electron-rebuild": "^1.10.1",
-    "eslint": "^7.6.0",
-    "eslint-plugin-import": "^2.21.1",
-    "file-loader": "^5.0.2",
-    "graceful-fs": "^4.2.4",
-    "html-loader": "0.5.5",
+    "electron-rebuild": "^2.3.5",
+    "eslint": "^7.31.0",
+    "eslint-plugin-import": "^2.23.4",
+    "file-loader": "^6.2.0",
+    "graceful-fs": "^4.2.6",
+    "html-loader": "2.1.2",
     "json-loader": "0.5.7",
-    "node-abi": "^2.18.0",
-    "node-gyp": "^7.0.0",
-    "node-sass": "^4.14.1",
-    "npmlog": "4.1.2",
-    "npx": "^10.2.0",
-    "pug": "^2.0.4",
+    "lru-cache": "^6.0.0",
+    "macos-release": "^2.5.0",
+    "ngx-sortablejs": "^11.1.0",
+    "ngx-toastr": "^14.0.0",
+    "node-abi": "^2.30.0",
+    "node-sass": "^6.0.1",
+    "npmlog": "5.0.0",
+    "npx": "^10.2.2",
+    "patch-package": "^6.4.7",
+    "pug": "^3.0.2",
     "pug-html-loader": "1.1.5",
     "pug-lint": "^2.6.0",
     "pug-loader": "^2.4.0",
     "pug-static-loader": "2.0.0",
-    "raw-loader": "4.0.1",
-    "sass-loader": "^8.0.0",
+    "raw-loader": "4.0.2",
+    "sass-loader": "^12.1.0",
+    "shell-quote": "^1.7.2",
     "shelljs": "0.8.4",
-    "source-code-pro": "^2.30.2",
+    "slugify": "^1.5.3",
+    "sortablejs": "^1.14.0",
+    "source-code-pro": "^2.38.0",
+    "source-map-loader": "^3.0.0",
     "source-sans-pro": "3.6.0",
-    "style-loader": "^1.1.4",
-    "svg-inline-loader": "^0.8.0",
-    "to-string-loader": "1.1.6",
-    "tslib": "^2.0.0",
-    "typedoc": "^0.18.0",
-    "typescript": "^3.9.3",
-    "url-loader": "^3.0.0",
-    "val-loader": "2.1.1",
-    "webpack": "^5.0.0-beta.18",
-    "webpack-cli": "^3.3.12",
-    "yaml-loader": "0.6.0"
+    "ssh2": "^1.1.0",
+    "style-loader": "^3.1.0",
+    "svg-inline-loader": "^0.8.2",
+    "ts-loader": "^9.2.3",
+    "tslib": "^2.3.0",
+    "typedoc": "^0.21.4",
+    "typescript": "^4.3.5",
+    "url-loader": "^4.1.1",
+    "val-loader": "4.0.0",
+    "webpack": "^5.46.0",
+    "webpack-bundle-analyzer": "^4.4.2",
+    "webpack-cli": "^4.7.0",
+    "yaml-loader": "0.6.0",
+    "zone.js": "^0.11.4"
   },
   "resolutions": {
-    "*/node-abi": "^2.14.0"
+    "lzma-native": "^8.0.0",
+    "*/node-abi": "^2.30.0",
+    "**/graceful-fs": "^4.2.4"
   },
   "scripts": {
-    "build": "npm run build:typings && webpack --color --config app/webpack.main.config.js && webpack --color --config app/webpack.config.js && webpack --color --config terminus-core/webpack.config.js && webpack --color --config terminus-settings/webpack.config.js && webpack --color --config terminus-terminal/webpack.config.js && webpack --color --config terminus-plugin-manager/webpack.config.js && webpack --color --config terminus-community-color-schemes/webpack.config.js && webpack --color --config terminus-ssh/webpack.config.js && webpack --color --config terminus-serial/webpack.config.js && webpack --color --config terminus-theme-hype/webpack.config.js",
+    "build": "npm run build:typings && node scripts/build-modules.js",
     "build:typings": "node scripts/build-typings.js",
-    "watch": "cross-env TERMINUS_DEV=1 webpack --progress --color --watch",
-    "start": "cross-env TERMINUS_DEV=1 electron app --debug",
+    "watch": "cross-env TABBY_DEV=1 webpack --progress --color --watch",
+    "start": "cross-env TABBY_DEV=1 electron app --debug --inspect",
     "start:prod": "electron app --debug",
-    "prod": "cross-env TERMINUS_DEV=1 electron app",
-    "docs": "typedoc --out docs/api terminus-core/src && typedoc --out docs/api/terminal --tsconfig terminus-terminal/tsconfig.typings.json terminus-terminal/src && typedoc --out docs/api/settings --tsconfig terminus-settings/tsconfig.typings.json terminus-settings/src",
+    "prod": "cross-env TABBY_DEV=1 electron app",
+    "docs": "typedoc --out docs/api --tsconfig tabby-core/src/tsconfig.typings.json tabby-core/src/index.ts && typedoc --out docs/api/terminal --tsconfig tabby-terminal/tsconfig.typings.json tabby-terminal/src/index.ts && typedoc --out docs/api/local --tsconfig tabby-local/tsconfig.typings.json tabby-local/src/index.ts && typedoc --out docs/api/settings --tsconfig tabby-settings/tsconfig.typings.json tabby-settings/src/index.ts",
     "lint": "eslint --ext ts */src */lib",
-    "postinstall": "node ./scripts/install-deps.js"
+    "postinstall": "node ./scripts/install-deps.js",
+    "patch": "patch-package; cd web; patch-package"
   },
-  "repository": "eugeny/terminus",
-  "author": "Eugene Pankov",
-  "license": "MIT"
+  "private": true
 }
