@@ -5,12 +5,12 @@ export function parseArgs (argv: string[], cwd: string): any {
         argv = argv.slice(1)
     }
 
-    return require('yargs')
-        .usage('terminus [command] [arguments]')
+    return require('yargs/yargs')(argv.slice(1))
+        .usage('tabby [command] [arguments]')
         .command('open [directory]', 'open a shell in a directory', {
             directory: { type: 'string', 'default': cwd },
         })
-        .command('run [command...]', 'run a command in the terminal', {
+        .command(['run [command...]', '/k'], 'run a command in the terminal', {
             command: { type: 'string' },
         })
         .command('profile [profileName]', 'open a tab with specified profile', {
@@ -41,5 +41,5 @@ export function parseArgs (argv: string[], cwd: string): any {
             type: 'boolean',
         })
         .help('help')
-        .parse(argv.slice(1))
+        .parse()
 }
