@@ -1,6 +1,6 @@
-import * as createLRU from 'lru-cache'
+import * as LRU from 'lru-cache'
 import * as fs from 'fs'
-const lru = createLRU({ max: 256, maxAge: 250 })
+const lru = new LRU({ max: 256, maxAge: 250 })
 const origLstat = fs.realpathSync.bind(fs)
 
 // NB: The biggest offender of thrashing realpathSync is the node module system
