@@ -1,89 +1,53 @@
 import * as fs from 'mz/fs'
 import * as path from 'path'
+import * as remote from '@electron/remote'
+import { PluginInfo } from '../../tabby-core/src/api/mainProcess'
+
 const nodeModule = require('module') // eslint-disable-line @typescript-eslint/no-var-requires
-const nodeRequire = (global as any).require
 
-function normalizePath (path: string): string {
+const nodeRequire = global['require']
+
+function normalizePath (p: string): string {
     const cygwinPrefix = '/cygdrive/'
-    if (path.startsWith(cygwinPrefix)) {
-        path = path.substring(cygwinPrefix.length).replace('/', '\\')
-        path = path[0] + ':' + path.substring(1)
+    if (p.startsWith(cygwinPrefix)) {
+        p = p.substring(cygwinPrefix.length).replace('/', '\\')
+        p = p[0] + ':' + p.substring(1)
     }
-    return path
+    return p
 }
 
-global['module'].paths.map((x: string) => nodeModule.globalPaths.push(normalizePath(x)))
+const builtinPluginsPath = process.env.TABBY_DEV ? path.dirname(remote.app.getAppPath()) : path.join((process as any).resourcesPath, 'builtin-plugins')
 
-if (process.env.TERMINUS_DEV) {
-    nodeModule.globalPaths.unshift(path.dirname(require('electron').remote.app.getAppPath()))
-}
-
-const builtinPluginsPath = process.env.TERMINUS_DEV ? path.dirname(require('electron').remote.app.getAppPath()) : path.join((process as any).resourcesPath, 'builtin-plugins')
-
-const userPluginsPath = path.join(
-    require('electron').remote.app.getPath('userData'),
-    'plugins',
-)
-
-if (!fs.existsSync(userPluginsPath)) {
-    fs.mkdir(userPluginsPath)
-}
-
-Object.assign(window, { builtinPluginsPath, userPluginsPath })
-nodeModule.globalPaths.unshift(builtinPluginsPath)
-nodeModule.globalPaths.unshift(path.join(userPluginsPath, 'node_modules'))
-// nodeModule.globalPaths.unshift(path.join((process as any).resourcesPath, 'app.asar', 'node_modules'))
-if (process.env.TERMINUS_PLUGINS) {
-    process.env.TERMINUS_PLUGINS.split(':').map(x => nodeModule.globalPaths.push(normalizePath(x)))
-}
-
-export type ProgressCallback = (current: number, total: number) => void // eslint-disable-line @typescript-eslint/no-type-alias
-
-export interface PluginInfo {
-    name: string
-    description: string
-    packageName: string
-    isBuiltin: boolean
-    version: string
-    author: string
-    homepage?: string
-    path?: string
-    info?: any
+const cachedBuiltinModules = {
+    '@angular/animations': require('@angular/animations'),
+    '@angular/common': require('@angular/common'),
+    '@angular/compiler': require('@angular/compiler'),
+    '@angular/core': require('@angular/core'),
+    '@angular/forms': require('@angular/forms'),
+    '@angular/platform-browser': require('@angular/platform-browser'),
+    '@angular/platform-browser/animations': require('@angular/platform-browser/animations'),
+    '@angular/platform-browser-dynamic': require('@angular/platform-browser-dynamic'),
+    '@ng-bootstrap/ng-bootstrap': require('@ng-bootstrap/ng-bootstrap'),
+    'ngx-toastr': require('ngx-toastr'),
+    rxjs: require('rxjs'),
+    'rxjs/operators': require('rxjs/operators'),
+    'zone.js/dist/zone.js': require('zone.js/dist/zone.js'),
 }
 
 const builtinModules = [
-    '@angular/animations',
-    '@angular/common',
-    '@angular/compiler',
-    '@angular/core',
-    '@angular/forms',
-    '@angular/platform-browser',
-    '@angular/platform-browser-dynamic',
-    '@ng-bootstrap/ng-bootstrap',
-    'ngx-toastr',
-    'rxjs',
-    'rxjs/operators',
-    'rxjs-compat/Subject',
-    'terminus-core',
-    'terminus-settings',
-    'terminus-terminal',
-    'zone.js/dist/zone.js',
+    ...Object.keys(cachedBuiltinModules),
+    'tabby-core',
+    'tabby-local',
+    'tabby-settings',
+    'tabby-terminal',
 ]
-
-const cachedBuiltinModules = {}
-builtinModules.forEach(m => {
-    const label = 'Caching ' + m
-    console.time(label)
-    cachedBuiltinModules[m] = nodeRequire(m)
-    console.timeEnd(label)
-})
 
 const originalRequire = (global as any).require
 ;(global as any).require = function (query: string) {
     if (cachedBuiltinModules[query]) {
         return cachedBuiltinModules[query]
     }
-    return originalRequire.apply(this, arguments)
+    return originalRequire.apply(this, [query])
 }
 
 const originalModuleRequire = nodeModule.prototype.require
@@ -94,13 +58,45 @@ nodeModule.prototype.require = function (query: string) {
     return originalModuleRequire.call(this, query)
 }
 
+export type ProgressCallback = (current: number, total: number) => void // eslint-disable-line @typescript-eslint/no-type-alias
+
+export function initModuleLookup (userPluginsPath: string): void {
+    global['module'].paths.map((x: string) => nodeModule.globalPaths.push(normalizePath(x)))
+
+    nodeModule.globalPaths.unshift(path.join(userPluginsPath, 'node_modules'))
+
+    if (process.env.TABBY_DEV) {
+        nodeModule.globalPaths.unshift(path.dirname(remote.app.getAppPath()))
+    }
+
+    nodeModule.globalPaths.unshift(builtinPluginsPath)
+    // nodeModule.globalPaths.unshift(path.join((process as any).resourcesPath, 'app.asar', 'node_modules'))
+    if (process.env.TABBY_PLUGINS) {
+        process.env.TABBY_PLUGINS.split(':').map(x => nodeModule.globalPaths.push(normalizePath(x)))
+    }
+
+    builtinModules.forEach(m => {
+        if (!cachedBuiltinModules[m]) {
+            cachedBuiltinModules[m] = nodeRequire(m)
+        }
+    })
+}
+
 export async function findPlugins (): Promise<PluginInfo[]> {
     const paths = nodeModule.globalPaths
     let foundPlugins: PluginInfo[] = []
     const candidateLocations: { pluginDir: string, packageName: string }[] = []
-    const PREFIX = 'terminus-'
+    const PREFIX = 'tabby-'
+    const LEGACY_PREFIX = 'terminus-'
+
+    const processedPaths = []
 
     for (let pluginDir of paths) {
+        if (processedPaths.includes(pluginDir)) {
+            continue
+        }
+        processedPaths.push(pluginDir)
+
         pluginDir = normalizePath(pluginDir)
         if (!await fs.exists(pluginDir)) {
             continue
@@ -113,7 +109,7 @@ export async function findPlugins (): Promise<PluginInfo[]> {
             })
         }
         for (const packageName of pluginNames) {
-            if (packageName.startsWith(PREFIX)) {
+            if (packageName.startsWith(PREFIX) || packageName.startsWith(LEGACY_PREFIX)) {
                 candidateLocations.push({ pluginDir, packageName })
             }
         }
@@ -126,16 +122,28 @@ export async function findPlugins (): Promise<PluginInfo[]> {
             continue
         }
 
-        const name = packageName.substring(PREFIX.length)
+        const name = packageName.startsWith(PREFIX) ? packageName.substring(PREFIX.length) : packageName.substring(LEGACY_PREFIX.length)
 
-        if (foundPlugins.some(x => x.name === name)) {
-            console.info(`Plugin ${packageName} already exists, overriding`)
-            foundPlugins = foundPlugins.filter(x => x.name !== name)
+        if (builtinModules.includes(packageName) && pluginDir !== builtinPluginsPath) {
+            continue
+        }
+
+        console.log(`Found ${name} in ${pluginDir}`)
+
+        const existing = foundPlugins.find(x => x.name === name)
+        if (existing) {
+            if (existing.isLegacy) {
+                console.info(`Plugin ${packageName} already exists, overriding`)
+                foundPlugins = foundPlugins.filter(x => x.name !== name)
+            } else {
+                console.info(`Plugin ${packageName} already exists, skipping`)
+                continue
+            }
         }
 
         try {
             const info = JSON.parse(await fs.readFile(infoPath, { encoding: 'utf-8' }))
-            if (!info.keywords || !(info.keywords.includes('terminus-plugin') || info.keywords.includes('terminus-builtin-plugin'))) {
+            if (!info.keywords || !(info.keywords.includes('terminus-plugin') || info.keywords.includes('terminus-builtin-plugin') || info.keywords.includes('tabby-plugin') || info.keywords.includes('tabby-builtin-plugin'))) {
                 continue
             }
             let author = info.author
@@ -144,6 +152,7 @@ export async function findPlugins (): Promise<PluginInfo[]> {
                 name: name,
                 packageName: packageName,
                 isBuiltin: pluginDir === builtinPluginsPath,
+                isLegacy: info.keywords.includes('terminus-plugin') || info.keywords.includes('terminus-builtin-plugin'),
                 version: info.version,
                 description: info.description,
                 author,
@@ -156,8 +165,7 @@ export async function findPlugins (): Promise<PluginInfo[]> {
     }
 
     foundPlugins.sort((a, b) => a.name > b.name ? 1 : -1)
-
-    ;(window as any).installedPlugins = foundPlugins
+    foundPlugins.sort((a, b) => a.isBuiltin < b.isBuiltin ? 1 : -1)
     return foundPlugins
 }
 
@@ -169,14 +177,14 @@ export async function loadPlugins (foundPlugins: PluginInfo[], progress: Progres
         console.info(`Loading ${foundPlugin.name}: ${nodeRequire.resolve(foundPlugin.path)}`)
         progress(index, foundPlugins.length)
         try {
-            const label = 'Loading ' + foundPlugin.name
-            console.time(label)
             const packageModule = nodeRequire(foundPlugin.path)
+            if (foundPlugin.packageName.startsWith('tabby-')) {
+                cachedBuiltinModules[foundPlugin.packageName.replace('tabby-', 'terminus-')] = packageModule
+            }
             const pluginModule = packageModule.default.forRoot ? packageModule.default.forRoot() : packageModule.default
-            pluginModule['pluginName'] = foundPlugin.name
-            pluginModule['bootstrap'] = packageModule.bootstrap
+            pluginModule.pluginName = foundPlugin.name
+            pluginModule.bootstrap = packageModule.bootstrap
             plugins.push(pluginModule)
-            console.timeEnd(label)
             await new Promise(x => setTimeout(x, 50))
         } catch (error) {
             console.error(`Could not load ${foundPlugin.name}:`, error)
