@@ -3,11 +3,10 @@ const builder = require('electron-builder').build
 const vars = require('./vars')
 
 const isTag = (process.env.GITHUB_REF || '').startsWith('refs/tags/')
-const isCI = !!process.env.GITHUB_REF
 
 builder({
     dir: true,
-    linux: ['deb', 'tar.gz', 'rpm'],
+    linux: ['deb', 'tar.gz', 'rpm', 'pacman'],
     config: {
         extraMetadata: {
             version: vars.version,
