@@ -4,10 +4,11 @@ const path = require('path')
 const vars = require('./vars')
 
 let lifecycles = []
-for (let dir of ['app', 'terminus-core', 'terminus-ssh', 'terminus-terminal']) {
+for (let dir of ['app', 'tabby-core', 'tabby-local', 'tabby-ssh', 'tabby-terminal']) {
     const build = rebuild({
         buildPath: path.resolve(__dirname, '../' + dir),
         electronVersion: vars.electronVersion,
+        arch: process.env.ARCH ?? process.arch,
         force: true,
     })
     build.catch(e => {
