@@ -11,14 +11,24 @@ sh.mkdir('-p', target)
 fs.writeFileSync(path.join(target, 'package.json'), '{}')
 sh.cd(target)
 vars.builtinPlugins.forEach(plugin => {
+  if (plugin === 'tabby-web') {
+    return
+  }
   log.info('install', plugin)
   sh.cp('-r', path.join('..', plugin), '.')
   sh.rm('-rf', path.join(plugin, 'node_modules'))
   sh.cd(plugin)
-  sh.exec(`npm install --only=prod`)
+  sh.exec(`yarn install --force --production`)
+
+
   log.info('rebuild', 'native')
   if (fs.existsSync('node_modules')) {
-    rebuild(path.resolve('.'), vars.electronVersion, process.arch, [], true)
+    rebuild({
+      buildPath: path.resolve('.'),
+      electronVersion: vars.electronVersion,
+      arch: process.env.ARCH ?? process.arch,
+      force: true,
+    })
   }
   sh.cd('..')
 })
