@@ -1,7 +1,4 @@
 module.exports = {
-  ignoreCompilerErrors: true,
-  excludeNotExported: true,
   excludePrivate: true,
-  excludeExternals: true,
-  mode: 'file'
+  excludeExternals: true
 }
